+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v2.15.0
 
 ### Added
 
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    break when updating to the latest JS SDK version
  * `ContractCreateFlow.executeWithSigner()` - Execute a flow using a signer
  * `AccountInfoFlow.verify[Signature|Transaction]WithSigner()` - Execute a flow using a signer
+ * A third party time source will be used for transaction ID generation when possible.
 
 ## v2.14.2